==> R/14a_replace_table.R <==


replace_table <- readr::read_rds("replace_table.rds")

replace_table_stage <- unique(tibble::tribble(
  ~wrong, ~right,
  "experienceing","experiencing",
  "extentions","extensions",
  "fiances","finances",
  "authoritive","authoritative",
  "pereceived","perceived",
  "crimial","criminal",
  "authorative","authoritative ",
  "sedentry","sedentary",
  "merclant","merchant",
  "veterinerary","veterinary",
  "stimates","estimates",
  "perisistent","persistent",
  "widdecombe","Widdecombe",
  "appologise","apologise",
  "resiliant","resilient",
  "avacuation","evacuation",
  "suggestins","suggesting",
  "relevation","revelation",
  "beaurocratic","bureaucratic ",
  "taliking","talking",
  "axieties","anxieties",
  "schemes's","schemes’",
  "urcontrolled","uncontrolled",
  "infonnation","information",
  "allegiations","allegations",
  "importart","important",
  "areaa","area",
  "plcs","places",
  "successfuly","successfully",
  "transfernng","transferring",
  "forumula","formula",
  "alocholic","alcoholic",
  "wereallowed","were allowed",
  "stevens","Stevens",
  "borrowiing","borrowing",
  "reevaluation","re-evaluation",
  "bureucracy","bureaucracy",
  "dicated","dictated",
  "donorAID","donor AID",
  "patientsto","patients to",
  "subsquently","subsequently",
  "chromosonal","chromosomal",
  "atmostpheres","atmospheres",
  "fpr","for",
  "restaints","restraints",
  "agandas","agendas",
  "filofaxes","Filofaxes",
  "got'em","got' em",
  "importnat","important",
  "widly","wildly",
  "manoevre","manoeuvre",
  "underpoliced","under policed",
  "availabl","available",
  "devicies","devices",
  "throrough","thorough",
  "inibited","inhibited",
  "assistace","assistance",
  "commisar","commissar",
  "embarrassement","embarrassment",
  "anying","anything",
  "whih","which",
  "nearsighted","near-sighted",
  "dependencey","dependency",
  "freephone","Freephone",
  "enteprise","enterprise",
  "contraverted","controverted",
  "mantlepiece","mantelpiece",
  "pressurre","pressure",
  "abslolute","absolute",
  "incenitive","incentive",
  "organsiational","organisational",
  "governors's","governors’",
  "opionion","opinion",
  "atempts","attempts",
  "extensificationAddendum","extensification Addendum",
  "daysavailable","days available",
  "circumtances","circumstances",
  "presidental","presidential",
  "satisfly","satisfy",
  "rembered","remembered",
  "semitism","Semitism",
  "manyof","many of",
  "buinesses","businesses",
  "lackof","lack of",
  "developeed","developed",
  "bedouin","Bedouin",
  "calummy","calumny",
  "recommendaton","recommendation",
  "exependiture","expenditure",
  "greif","grief",
  "plublic","public",
  "cathederal","cathedral",
  "creches","crèches",
  "creche","crèche",
  "governers","governors",
  "london","London",
  "carefull","careful",
  "faiis","fais",
  "substracted","subtracted",
  "toearnings","to earnings",
  "anybodys's","anybody’s",
  "shootand","shoot and",
  "descision","decision",
  "misconcieved","misconceived",
  "privlieged","privileged",
  "excutives","executives",
  "itpay","it pay",
  "succeessive","successive",
  "insiduous","insidious",
  "progremme","programme",
  "serverely","severely",
  "hydrid","hybrid",
  "demostration","demonstration",
  "politial","political",
  "understanably","understandably",
  "nationaliation","nationalisation",
  "amost","almost",
  "alsolutely","absolutely",
  "sedentory","sedentary",
  "sympatheticaly","sympathetically",
  "paitents","patients",
  "funiture","furniture",
  "perfectable","perfectible",
  "inprovement","improvement",
  "wuestions","questions",
  "accommadation","accommodation",
  "subisidies","subsidies",
  "concluson","conclusion",
  "considerble","considerable",
  "srrangements","arrangements",
  "anaesthesics","anaesthetics",
  "cosider","consider",
  "wihin","within",
  "inflexibilityof","inflexibility of",
  "weakess","weakness",
  "advantge","advantage",
  "precepitately","precipitately",
  "celebratong","celebrating",
  "presccription","prescription",
  "semitic","Semitic",
  "ironycc","irony",
  "continuned","continued",
  "spondored","sponsored",
  "destablising","destabilising",
  "proporton","proportion",
  "skeptical","sceptical",
  "catastrophy","catastrophe",
  "obsolesence","obsolescence",
  "fficial","official",
  "priod","period",
  "doctors's","doctors’",
  "compoundng","compounding",
  "injusticies","injustices",
  "dimissal","dismissal",
  "tranparent","transparent",
  "considable","considerable",
  "progaganda","propaganda",
  "sectariansism","sectarianism",
  "enablle","enable",
  "familes","families",
  "prorportion","proportion",
  "negotiatiing","negotiating",
  "controlexpenditure","control expenditure",
  "euphemestically","euphemistically",
  "competititive","competitive",
  "relacement","replacement",
  "improvemens","improvements",
  "nucler","nuclear",
  "outdatted","outdated",
  "uselees","useless",
  "amphibous","amphibious",
  "cauutionary","cautionary",
  "safeguardss","safeguards",
  "lessone","lesson",
  "parients","parents",
  "represetations","representations",
  "domesday","Domesday",
  "throughtout","throughout",
  "multideparimental","multi-departmental",
  "changeabout","change about",
  "undertstand","understand",
  "sould","should",
  "nonpolluting","non-polluting",
  "contracters","contractors",
  "ophthmalogy","ophthalmology",
  "methodist","Methodist",
  "riviera","Riviera",
  "supplementey","supplementary",
  "straightlaced","straight-laced",
  "threfore","therefore",
  "northwestern","north-western",
  "peviously","previously",
  "cemetaries","cemeteries",
  "underly","underlie",
  "remoued","removed",
  "maiestic","majestic",
  "straingers","strangers",
  "transportacion","transportation",
  "shoold","should",
  "unbelievalbe","unbelievable",
  "accontants","accountants",
  "misuderstood","misunderstood",
  "oganisation","organisation",
  "rottweilers","Rottweilers",
  "rottweiler","Rottweiler",
  "symathetically","sympathetically",
  "cloudcuckoo","cloud cuckoo",
  "invidous","invidious",
  "leukemia","leukaemia",
  "tumor","tumour",
  "injuiries","injuries",
  "zerorating","zero rating",
  "approximisation","approximation",
  "buscuits","biscuits",
  "fancyfree","fancy-free",
  "exploition","exploitation",
  "neighour","neighbour",
  "mconstituencies","in constituencies",
  "alredy","already",
  "embarassed","embarrassed",
  "experiemental","experimental",
  "structue","structure",
  "ecnomically","economically",
  "vacanices","vacancies",
  "shool","school",
  "microSieverts","micro-Sieverts",
  "resum√©","resumé",
  "willigness","willingness",
  "industralists","industrialists",
  "electricial","electrical",
  "olympics","Olympics",
  "erradicated","eradicated",
  "valution","valuation",
  "extravance","extravgaance",
  "expnditure","expenditure",
  "claustrophic","claustrophobic",
  "poisition","position",
  "indepdent","independent",
  "synethetic","synthetic",
  "prositively","positively",
  "circumstances√§¬∫","circumstances",
  "wealthly","wealthy",
  "passerby","passer-by",
  "veiew","view",
  "nterruption","interruption"
))

replace_table_stage <- replace_table_stage[!(replace_table_stage$wrong %in% replace_table$wrong),]

replace_table <- unique(dplyr::bind_rows(replace_table, replace_table_stage))

readr::write_rds(replace_table, "replace_table.rds")
